Use "_Atomic TYPE" instead of "_Atomic(TYPE)" to help CLion C parser.
CLion fails to parse "_Atomic(TYPE)" and deduces two different
declarations for TYPE. For instance, `extern _Atomic(size_t) var;`
is detected as a re-declaration of `size_t` (which is a typedef)
and any code using `size_t` is marked with a syntax error.

This only happens when using `_Atomic(TYPE)` in a `typedef` or
an `extern` statement. Since `_Atomic(TYPE)` and `_Atomic TYPE`
are equivalent in this case, we replace some occurrences with
`_ATOMIC(TYPE)` in header files.

Note that CLion will *still* fail to parse it correctly but this will
no more pollute other code, thereby improving (a bit) the programmer's
life. Unfortunately, this has been a long-standing bug for CLion which
we hope will one day be resolved.